Add files via upload
This is a learning management system based on php (html) and css and js for framework , i used python flask lib. for execution. has user panel along with admin panel . my project is basically a platform to learn thing by watching course videos ,its is an very good platform for both students as well as teachers , admin can uploads the course and lessons to the course. for execution of this program you'll required python with flask , xampp application with my sql and apache , php database which is provided in the program it self u just have to import.